+++ dolby_vision/CHANGELOG.md
@@ -1,6 +1,7 @@
 ## ??
 
 - Add support for compressed RPU format.
+- Add `RpuDataHeader::nlq_pred_pivot_value` field, parsed for profile 4 and 7.
 
 ## 1.6.2